requirements.txt
Flask==3.0.0
flask-mysqldb==1.0.1
mysqlclient==2.2.0
